// src/components/sign/api/sign.d.ts
export type ServiceAgreementStatus = 1 | 2
// ...
export interface CustomerInfoDataForUI {
  id: number | null
  status: ServiceAgreementStatus
  companyName: string
  industry: string | null
  companyArea: string
  companyAddress: string
  liaisonName: string
  liaisonPosition: string
  liaisonPhone: string
  yearUsableCharge: number | null
  isTimeOfUsePricingEnabled: boolean
  peakPercentage: number | null
  superPeakPercentage: number | null
  standardPercentage: number | null
  valleyPercentage: number | null
  comment: string | null
}